refactor: lift sessions store to HttpAdapter; identity routes now accept session cookie and configured flag
